skus/requirements.txt
asgiref==3.2.10
attrs==20.2.0
Django==3.1.2
django-admin==2.0.1
django-cors-headers==3.5.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
factory-boy==3.1.0
Faker==4.14.0
importlib-metadata==2.0.0
iniconfig==1.0.0
mysqlclient==2.0.1
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.1
pytest-django==3.10.0
python-dateutil==2.8.1
pytz==2020.1
screen==1.0.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
xlwt==1.3.0
zipp==3.3.0
